Add files via upload
This is the initial design I created for the Sand Paper Bot(I believe that's the name but it's essentially the bot that wasn't designed yet). This is only the rough draft of how it will look-the measurements are sort of random right now. So only think of this as where you like the looks of it or not. The 8 holes on the sides are for the ultrasonic sensor(HC-SR04, the cheap Arduino one)